+++ src/variaveis.cpp
@@ -5,7 +5,7 @@ void Variaveis::setIndice(int i){
     this->indice = i;
 }
 
-void Variaveis::setCoeficiente(double c){
+void Variaveis::setCoeficiente(float c){
     this->coeficiente = c;
 }
 
@@ -13,6 +13,6 @@ int Variaveis::getIndice(){
     return this->indice;
 }
 
-double Variaveis::getCoeficiente(){
+float Variaveis::getCoeficiente(){
     return this->coeficiente;
 }
